+++ config_baselines.tf
@@ -28,7 +28,7 @@ POLICY
 data "aws_iam_policy_document" "recoder_publish_policy" {
   statement {
     actions = ["s3:PutObject"]
-    resources = ["${module.audit_log_bucket.this_bucket.arn}/config/AWSLogs/${var.aws_account_id}/*"]
+    resources = ["${local.audit_log_bucket_arn}/config/AWSLogs/${var.aws_account_id}/*"]
 
     condition {
       test = "StringLike"
@@ -39,7 +39,7 @@ data "aws_iam_policy_document" "recoder_publish_policy" {
 
   statement {
     actions = ["s3:GetBucketAcl"]
-    resources = [module.audit_log_bucket.this_bucket.arn]
+    resources = [local.audit_log_bucket_arn]
   }
 
   statement {
@@ -85,7 +85,7 @@ resource "aws_iam_role_policy_attachment" "recoder_read_policy" {
 module "config_baseline_ap-northeast-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -98,7 +98,7 @@ module "config_baseline_ap-northeast-1" {
 module "config_baseline_ap-northeast-2" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -111,7 +111,7 @@ module "config_baseline_ap-northeast-2" {
 module "config_baseline_ap-south-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -124,7 +124,7 @@ module "config_baseline_ap-south-1" {
 module "config_baseline_ap-southeast-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -137,7 +137,7 @@ module "config_baseline_ap-southeast-1" {
 module "config_baseline_ap-southeast-2" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -150,7 +150,7 @@ module "config_baseline_ap-southeast-2" {
 module "config_baseline_ca-central-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -163,7 +163,7 @@ module "config_baseline_ca-central-1" {
 module "config_baseline_eu-central-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -176,7 +176,7 @@ module "config_baseline_eu-central-1" {
 module "config_baseline_eu-north-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -189,7 +189,7 @@ module "config_baseline_eu-north-1" {
 module "config_baseline_eu-west-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -202,7 +202,7 @@ module "config_baseline_eu-west-1" {
 module "config_baseline_eu-west-2" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -215,7 +215,7 @@ module "config_baseline_eu-west-2" {
 module "config_baseline_eu-west-3" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -228,7 +228,7 @@ module "config_baseline_eu-west-3" {
 module "config_baseline_sa-east-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -241,7 +241,7 @@ module "config_baseline_sa-east-1" {
 module "config_baseline_us-east-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -254,7 +254,7 @@ module "config_baseline_us-east-1" {
 module "config_baseline_us-east-2" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -267,7 +267,7 @@ module "config_baseline_us-east-2" {
 module "config_baseline_us-west-1" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name
@@ -280,7 +280,7 @@ module "config_baseline_us-west-1" {
 module "config_baseline_us-west-2" {
   source = "./modules/config-baseline"
   iam_role_arn = aws_iam_role.recorder.arn
-  s3_bucket_name = module.audit_log_bucket.this_bucket.id
+  s3_bucket_name = local.audit_log_bucket_id
   s3_key_prefix = var.config_s3_bucket_key_prefix
   delivery_frequency = var.config_delivery_frequency
   sns_topic_name = var.config_sns_topic_name