+++ examples/external-bucket/bucket.tf
@@ -1,0 +1,54 @@
+resource "aws_s3_bucket" "logs" {
+  bucket        = var.audit_s3_bucket_name
+  acl           = "private"
+  force_destroy = true
+}
+
+resource "aws_s3_bucket_policy" "logs" {
+  bucket = aws_s3_bucket.logs.id
+  policy = <<END_OF_POLICY
+{
+  "Version": "2012-10-17",
+  "Statement": [
+    {
+      "Sid": "AWSCloudTrailAclCheckForConfig",
+      "Effect": "Allow",
+      "Principal": {"Service": "config.amazonaws.com"},
+      "Action": "s3:GetBucketAcl",
+      "Resource": "${aws_s3_bucket.logs.arn}"
+    },
+    {
+      "Sid": " AWSCloudTrailWriteForConfig",
+      "Effect": "Allow",
+      "Principal": {"Service": "config.amazonaws.com"},
+      "Action": "s3:PutObject",
+      "Resource": "${aws_s3_bucket.logs.arn}/config/AWSLogs/${data.aws_caller_identity.current.account_id}/Config/*",
+      "Condition": {"StringEquals": {"s3:x-amz-acl": "bucket-owner-full-control"}}
+    },
+    {
+        "Sid": "AWSCloudTrailAclCheckForCloudTrail",
+        "Effect": "Allow",
+        "Principal": {
+            "Service": "cloudtrail.amazonaws.com"
+        },
+        "Action": "s3:GetBucketAcl",
+        "Resource": "${aws_s3_bucket.logs.arn}"
+    },
+    {
+        "Sid": "AWSCloudTrailWriteForCloudTrail",
+        "Effect": "Allow",
+        "Principal": {
+            "Service": "cloudtrail.amazonaws.com"
+        },
+        "Action": "s3:PutObject",
+        "Resource": "${aws_s3_bucket.logs.arn}/cloudtrail/AWSLogs/${data.aws_caller_identity.current.account_id}/*",
+        "Condition": {
+            "StringEquals": {
+                "s3:x-amz-acl": "bucket-owner-full-control"
+            }
+        }
+    }
+  ]
+}
+END_OF_POLICY
+}