+++ modules/secure-bucket/outputs.tf
@@ -1,9 +1,9 @@
 output "this_bucket" {
   description = "This S3 bucket."
-  value       = aws_s3_bucket.content
+  value       = var.enabled ? aws_s3_bucket.content[0] : null
 }
 
 output "log_bucket" {
   description = "The S3 bucket used for storing access logs of this bucket."
-  value       = aws_s3_bucket.access_log
+  value       = var.enabled ? aws_s3_bucket.access_log[0] : null
 }